% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltbt.R
\name{caf.mle}
\alias{caf.mle}
\title{Estimate the frequency of carriers of all functional rare variants in population. Maximum likelihood estimator (MLE) was used to account for the case/control ascertainment.}
\usage{
caf.mle(gamma.seq, Kprev.x, p0, p1, n.mut, n)
}
\arguments{
\item{gamma.seq}{a series of assumed rare-variant effect sizes (gamma)}

\item{Kprev.x}{the prevalence of disease in population}

\item{p0}{the ascertainment ratio for control, i.e., Pcontrol / (1 - K)}

\item{p1}{the ascertainment ratio for case, i.e., Pcase / K}

\item{n.mut}{the number of carriers in case/control}

\item{n}{the number of all subjects in case/control}
}
\value{
the most likely estimate of frequency of carriers in population calculated for each assumed gamma.seq
}
\description{
Estimate the frequency of carriers of all functional rare variants in population. Maximum likelihood estimator (MLE) was used to account for the case/control ascertainment.
}
