% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltbt.R
\name{ctrl.x}
\alias{ctrl.x}
\title{Calculate the likelihood of being a control given being a carrier with the specified rare-variant effect size.}
\usage{
ctrl.x(gamma.seq, genofreq, b, Kprev.x)
}
\arguments{
\item{gamma.seq}{a series of assumed rare-variant effect sizes (gamma)}

\item{genofreq}{the frequency of carriers in population}

\item{b}{the adjustment for effect size from the baseline gamma (1 if no adjustment neeed)}

\item{Kprev.x}{the prevalence of disease in population}
}
\value{
likelihood values calculated for each assumed gamma.seq
}
\description{
Calculate the likelihood of being a control given being a carrier with the specified rare-variant effect size.
}
