% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltbt.R
\name{lik.ctrl}
\alias{lik.ctrl}
\title{Calculate the likelihood of an individual being a control given being a carrier with the specified PRS score and rare-variant effect size.}
\usage{
lik.ctrl(gamma.seq, prs, b, Kprev.x, h2.prs.x)
}
\arguments{
\item{gamma.seq}{a series of assumed rare-variant effect sizes (gamma)}

\item{prs}{the polygenic risk score of carrier}

\item{b}{the adjustment for effect size from the baseline gamma (1 if no adjustment neeed)}

\item{Kprev.x}{the prevalence of disease in population}

\item{h2.prs.x}{the heritability explained by PRS (h2L[x])}
}
\value{
log likelihood values calculated for each assumed gamma.seq
}
\description{
Calculate the likelihood of an individual being a control given being a carrier with the specified PRS score and rare-variant effect size.
}
