% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltbt.R
\name{run.ltbt}
\alias{run.ltbt}
\title{Perform the Likelihood Ratio Test of liability threshold burden test (LTBT).}
\usage{
run.ltbt(gamma.seq, gt, prs, outcome, Kprev.x, h2.prs.x = NULL)
}
\arguments{
\item{gamma.seq}{a series of assumed rare-variant effect sizes (gamma)}

\item{gt}{a genotype vector for each individual in the cohort (0: non-carrier, 1: carrier)}

\item{prs}{a vector of polygenic risk scores for each individual in the cohort. It is assumed to have been standardized to the mean of 0 and variance of 1 in population.}

\item{outcome}{an outcome vector for each individual in the  cohort (0: control, 1: case)}

\item{Kprev.x}{the prevalence of disease in population}

\item{h2.prs.x}{the heritability explained by PRS (h2L[x]). It is estimated from cohort if it is left as NULL.}
}
\value{
LTBT test results
}
\description{
Perform the Likelihood Ratio Test of liability threshold burden test (LTBT).
}
